% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cor_CI.R
\name{add_cor_CI}
\alias{add_cor_CI}
\title{Calculate confidence intervals (CIs) for correlation estimates}
\usage{
add_cor_CI(r, n, alpha = 0.05, ci_offset = 3)
}
\arguments{
\item{r}{Correlation estimates.
Can be scalar or vector, each non-NA value must be between -1 and 1, inclusive.
Cases with NA \code{r} have all \code{NA} outputs (means and CIs).}

\item{n}{Sample sizes.
Can be scalar or vector.
If both \code{r} and \code{n} are vectors, they must have the same length.
Cases with NAs or with \code{n < ci_offset} have NA CIs.}

\item{alpha}{The complement of the confidence level (default 0.05 results in 1 - 0.05 = 95\% CIs).}

\item{ci_offset}{The amount by which the sample size gets reduced in the variance conversion formula.
For Normal data, Pearson should use 3 and CorSym 1.5, but different values may be better for other distributions.}
}
\value{
A matrix with a row for every input (single row if \code{r} and \code{n} are both scalars, as many rows as the lenghts of either \code{r} or \code{n} otherwise), and three named columns: \code{r} repeats the input, and \code{CIL} and \code{CIU} are the lower and upper CIs.
}
\description{
This function calculates CIs based on the Fisher transformation and theory originally derived for Pearson estimator, now validated empirically for CorSym.
Only the mean estimates and the sample sizes are needed for these calculations.
}
\examples{
# make up values for example
# a vector of pre-calculated correlation estimates
rs <- c( 0.3, 0.5, 0.9 )
# and sample sizes
ns <- c( 10, 100, 1000 )

# calculate CIs for these three cases:
out <- add_cor_CI( rs, ns )

}
\seealso{
\code{\link[=corsym]{corsym()}} and \code{\link[=pearson]{pearson()}} use this function internally.
}
