% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsym.R
\name{corsym}
\alias{corsym}
\title{Calculate CorSym estimates with confidence intervals (CIs)}
\usage{
corsym(x, y = NULL, alpha = 0.05, ci_offset = 1.5)
}
\arguments{
\item{x}{First vector of variables, or a 2-column matrix with the variables of interest}

\item{y}{Second vector of variables}

\item{alpha}{The complement of the confidence level (default 0.05 results in 1 - 0.05 = 95\% CIs).}

\item{ci_offset}{The amount by which the sample size gets reduced in the variance conversion formula.
For Normal data, Pearson should use 3 and CorSym 1.5, but different values may be better for other distributions.}
}
\value{
A vector with three values: the CorSym correlation estimate between \code{x} and \code{y}, and the lower and upper CIs.
}
\description{
This function estimates the correlation of two exchangeable variables without bias.
The output is deterministic, the same for all orders within individual pairs.
Pairs with any missingness are automatically ignored.
}
\examples{
# generate some random independent data, just for example
n <- 53
x <- rnorm( n )
y <- rnorm( n )
# calculate CorSym correlation with CIs
corsym( x, y )

}
