% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bias_test.R
\name{order_bias_test}
\alias{order_bias_test}
\title{Test for order bias}
\usage{
order_bias_test(data, f0 = 0.5)
}
\arguments{
\item{data}{A data.frame (including tibble) with at least two columns: \code{n} the sample size (number of pairs), and \code{x} the number of pairs in a given order.}

\item{f0}{The expected frequency of pairs in the given order, under the null hypothesis.}
}
\value{
The input data frame with two more columns: \code{f} is the sample order frequency (equal to \code{x/n}), and \code{pval} the two-sided Binomial p-value.
}
\description{
This function performs a test for whether data is appearing in a biased order or not.
The test calculates a two-sided Binomial p-value under the null hypothesis that the true frequency of the counted order is \code{f0} (0.5 by default).
}
\examples{
# toy data used in the test
data <- data.frame(
  n = c( 30, 50, 100 ),
  x = c( 12, 20,  13 )
)

# add frequencies and p-values!
data <- order_bias_test( data )

}
