% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{pearson}
\alias{pearson}
\title{Calculate Pearson estimates with confidence intervals (CIs)}
\usage{
pearson(x, y = NULL, alpha = 0.05, ci_offset = 3)
}
\arguments{
\item{x}{First vector of variables, or a 2-column matrix with the variables of interest}

\item{y}{Second vector of variables}

\item{alpha}{The complement of the confidence level (default 0.05 results in 1 - 0.05 = 95\% CIs).}

\item{ci_offset}{The amount by which the sample size gets reduced in the variance conversion formula.
For Normal data, Pearson should use 3 and CorSym 1.5, but different values may be better for other distributions.}
}
\value{
A vector with three values: the Pearson correlation estimate between \code{x} and \code{y}, and the lower and upper CIs.
}
\description{
This function is a simple wrapper around \code{\link[stats:cor]{stats::cor()}} with the same interface and missingness handling of \code{\link[=corsym]{corsym()}}, particularly it adds CIs to the estimates.
Pairs with any missingness are automatically ignored (matching \code{\link[stats:cor]{stats::cor()}} with non-default option \code{use = 'complete.obs'}).
Also, unlike \code{\link[stats:cor]{stats::cor()}}, this function accepts only two variables as inputs and does not calculate correlation matrices, just the correlation between the two variables.
Pearson is included in this package to constrast to \code{\link[=corsym]{corsym()}} in handling exchangeable variables.
}
\examples{
# generate some random independent data, just for example
n <- 53
x <- rnorm( n )
y <- rnorm( n )
# calculate Pearson correlation with CIs
pearson( x, y )

}
