% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_order.R
\name{randomize_order}
\alias{randomize_order}
\title{Randomize ordering of each row of a matrix}
\usage{
randomize_order(X)
}
\arguments{
\item{X}{Matrix of data.}
}
\value{
The matrix with randomized rows.
}
\description{
This function shuffles the values in each row, so they appear in a random order.
}
\examples{
# dummy data in order
X <- matrix( 1:20, nrow = 10, ncol = 2, byrow = TRUE )

# same data but with each row in random order
X_randomized <- randomize_order( X )

}
