% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsym-package.R
\docType{package}
\name{corsym-package}
\alias{corsym}
\alias{corsym-package}
\title{corsym: Correlation Estimation For Exchangeable/Symmetric Pairs}
\description{
We implement a new correlation estimator, CorSym, designed for exchangeable variables, where the ordering of the two values in the pair is arbitrary. This kind of data arises frequently in the study of assortative pairing (for example, income in a couple). The standard Pearson estimator is sensitive to such ordering and can be highly biased when the order is biased (when the first value tends to have lower or higher values than the second value). CorSym gives the same estimate deterministically for all orders within each pair, and estimates the desired correlation without bias (variables must be exchangeable). The package also includes utilities to simulate biased orders and test for order bias.
}
\author{
\strong{Maintainer}: Alejandro Ochoa \email{alejandro.ochoa@duke.edu} (\href{https://orcid.org/0000-0003-4928-3403}{ORCID})

Authors:
\itemize{
  \item Alejandro Ochoa \email{alejandro.ochoa@duke.edu} (\href{https://orcid.org/0000-0003-4928-3403}{ORCID})
  \item Gabriel Kennedy \email{gabriel.kennedy@duke.edu} (\href{https://orcid.org/0000-0003-4412-3777}{ORCID})
}

}
\keyword{internal}
