% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_order.R
\name{partial_order}
\alias{partial_order}
\title{Apply partial ordering to the rows of a matrix}
\usage{
partial_order(X, f = 1)
}
\arguments{
\item{X}{Matrix of data.}

\item{f}{Proportion of rows to force to have extreme order (sorted in increasing order).}
}
\value{
The input matrix with a random proportion \code{f} of rows sorted in increasing order, the rest equals the input data.
}
\description{
This function reorders a proportion of rows so values appear in increasing order.
}
\examples{
# dummy independent data in random order
X <- matrix( rnorm( 20 ), nrow = 10, ncol = 2 )

# same data but in extreme order by default
X_extreme <- partial_order( X )

# apply a partial ordering
X_partial <- partial_order( X, 0.5 )

}
